Refactor fragments.
<h1>Rationale:</h1>

This commit provides no functional changes.  It simply puts all User facing fragment classes into a separate directory in order to manage the chat classes for maintenance and testing.
