+++ app/src/main/java/com/pajato/android/gamechat/chat/fragment/ShowOfflineFragment.java
@@ -15,20 +15,25 @@
  * see http://www.gnu.org/licenses
  */
 
-package com.pajato.android.gamechat.chat;
+package com.pajato.android.gamechat.chat.fragment;
 
 import com.pajato.android.gamechat.R;
+import com.pajato.android.gamechat.chat.BaseChatFragment;
+import com.pajato.android.gamechat.event.ClickEvent;
 
-/**
- * Provide a fragment to handle the case where there are no joined rooms.
- *
- * @author Paul Michael Reilly
- */
-public class ShowNoJoinedRoomsFragment extends BaseChatFragment {
+import org.greenrobot.eventbus.Subscribe;
+
+
+public class ShowOfflineFragment extends BaseChatFragment {
 
     // Public instance methods.
 
-    /** Set the layout file. */
-    @Override public int getLayout() {return R.layout.fragment_chat_no_joined_rooms;}
+    /** Provide a placeholder subscriber to satisfy the event bus contract. */
+    @Subscribe public void onClick(final ClickEvent event) {
+        // Use a logging placeholder.
+        logEvent("onClick (showOffline)");
+    }
 
+    /** Establish the layout file to show that the app is offline due to network loss. */
+    @Override public int getLayout() {return R.layout.fragment_chat_offline;}
 }
