+++ app/src/main/java/com/pajato/android/gamechat/chat/fragment/ShowRoomListFragment.java
@@ -15,13 +15,15 @@
  * see http://www.gnu.org/licenses
  */
 
-package com.pajato.android.gamechat.chat;
+package com.pajato.android.gamechat.chat.fragment;
 
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.View;
 
 import com.pajato.android.gamechat.R;
+import com.pajato.android.gamechat.chat.BaseChatFragment;
+import com.pajato.android.gamechat.chat.ChatManager;
 import com.pajato.android.gamechat.common.FabManager;
 import com.pajato.android.gamechat.common.adapter.MenuEntry;
 import com.pajato.android.gamechat.database.DBUtils;
@@ -33,21 +35,22 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static com.pajato.android.gamechat.chat.ChatFragmentType.createGroup;
+import static com.pajato.android.gamechat.chat.ChatFragmentType.createRoom;
 import static com.pajato.android.gamechat.chat.ChatFragmentType.joinRoom;
 
 /**
- * Provide a fragment to handle the display of the groups available to the current user.  This is
- * the top level view in the chat hierarchy.  It shows all the joined groups and allows for drilling
- * into rooms and chats within those rooms.
+ * Provide a fragment to handle the display of the rooms available to the current user.  This is the
+ * penultimate view in the chat hierarchy when there is more than one group.  It shows all the
+ * joined rooms and allows for drilling into chats within those rooms.
  *
  * @author Paul Michael Reilly
  */
-public class ShowGroupListFragment extends BaseChatFragment {
+public class ShowRoomListFragment extends BaseChatFragment {
 
     // Public class constants.
 
     /** The lookup key for the FAB game home memu. */
-    public static final String CHAT_GROUP_FAM_KEY = "chatGroupFamKey";
+    public static final String CHAT_ROOM_FAM_KEY = "chatRoomFamKey";
 
     // Public instance methods.
 
@@ -66,6 +69,9 @@ public class ShowGroupListFragment extends BaseChatFragment {
             case R.string.CreateGroupMenuTitle:
                 ChatManager.instance.chainFragment(createGroup, getActivity(), mItem);
                 break;
+            case R.string.CreateRoomMenuTitle:
+                ChatManager.instance.chainFragment(createRoom, getActivity(), mItem);
+                break;
             case R.string.JoinRoomsMenuTitle:
                 ChatManager.instance.chainFragment(joinRoom, getActivity(), mItem);
                 break;
@@ -75,40 +81,37 @@ public class ShowGroupListFragment extends BaseChatFragment {
         }
     }
 
-    /** Deal with the options menu by making the search button visible. */
+    /** Deal with the options menu creation by making the search item visible. */
     @Override public void onCreateOptionsMenu(final Menu menu, final MenuInflater inflater) {
-        // Turn on the search option.
+        // Turn on both the back and search buttons.
         setItemState(menu, R.id.search, true);
     }
 
     /** Initialize ... */
     @Override public void onInitialize() {
         super.onInitialize();
-        mItemListType = DBUtils.ChatListType.group;
+        mItemListType = DBUtils.ChatListType.room;
         initToolbar();
-        FabManager.chat.setMenu(CHAT_GROUP_FAM_KEY, getGroupMenu());
+        FabManager.chat.setMenu(CHAT_ROOM_FAM_KEY, getRoomMenu());
     }
 
-    /** Deal with the fragment's lifecycle by managing the progress bar and the FAB. */
+    /** Deal with the fragment's activity's lifecycle by managing the FAB. */
     @Override public void onResume() {
-        // Set the titles in the toolbar to the app title only; ensure that the FAB is visible, the
+        // Set the titles in the toolbar to the group name only; ensure that the FAB is visible, the
         // FAM is not and the FAM is set to the home chat menu; initialize the ad view; and set up
         // the group list display.
         super.onResume();
         FabManager.chat.setImage(R.drawable.ic_add_white_24dp);
         FabManager.chat.init(this);
         FabManager.chat.setVisibility(this, View.VISIBLE);
-        FabManager.chat.setMenu(this, CHAT_GROUP_FAM_KEY);
+        FabManager.chat.setMenu(this, CHAT_ROOM_FAM_KEY);
     }
 
-    // Private instance methods.
-
     /** Return the home FAM used in the top level show games and show no games fragments. */
-    private List<MenuEntry> getGroupMenu() {
+    private List<MenuEntry> getRoomMenu() {
         final List<MenuEntry> menu = new ArrayList<>();
         menu.add(getTintEntry(R.string.JoinRoomsMenuTitle, R.drawable.vd_casino_black_24px));
-        // TODO: add this when group selection is included:
-        //menu.add(getTintEntry(R.string.CreateRoomMenuTitle, R.drawable.vd_casino_black_24px));
+        menu.add(getTintEntry(R.string.CreateRoomMenuTitle, R.drawable.vd_casino_black_24px));
         menu.add(getTintEntry(R.string.CreateGroupMenuTitle, R.drawable.ic_group_add_black_24dp));
         return menu;
     }
